fix: int64 test migration batch 7 - update tests for NumPy 2.x arange int64 default
NumPy 2.x returns int64 from arange() by default. This batch updates tests
that used GetInt32/MakeGeneric<int> on arange-sourced arrays to use the
correct int64 accessors.

Test updates:
- Change GetInt32 -> GetInt64 for arange-sourced arrays
- Change MakeGeneric<int>() -> MakeGeneric<long>() for arange results
- Change np.array(new int[]) -> np.array(new long[]) where comparing to arange
- Fix NDIterator<int> -> NDIterator<long> for arange iteration
- Restore GetInt32 for explicit int32 arrays (np.array(42), np.array(new int[]))

Bug fix (Default.ClipNDArray.cs):
- Fixed mixed-dtype clip bug where int32 min/max arrays were read as int64
- Now casts min/max arrays to output dtype before calling SIMD kernel
- This prevented garbage values like 34359738376 (8 * 2^32+1)

Dtype preservation tests:
- Clip_Int32_PreservesDtype: use explicit int32 array
- Ravel_PreservesDtype_Int32: use explicit int32 array
- Reshape_Int32: use explicit int32 array
- Roll_PreservesDtype_Int32: use explicit int32 array

Test results: 103 failures -> 29 failures (74 tests fixed)

+++ test/NumSharp.UnitTest/Manipulation/NdArray.ReShape.Test.cs
@@ -14,26 +14,26 @@ namespace NumSharp.UnitTest.Manipulation
         {
             var nd = np.arange(6);
             var n1 = np.reshape(nd, 3, 2);
-            var n = n1.MakeGeneric<int>();
+            var n = n1.MakeGeneric<long>();
 
             Assert.IsTrue(n[0, 0] == 0);
             Assert.IsTrue(n[1, 1] == 3);
             Assert.IsTrue(n[2, 1] == 5);
 
-            n = np.reshape(np.arange(6), 2, 3, 1).MakeGeneric<int>();
+            n = np.reshape(np.arange(6), 2, 3, 1).MakeGeneric<long>();
             Assert.IsTrue(n[1, 1, 0] == 4);
             Assert.IsTrue(n[1, 2, 0] == 5);
 
-            n = np.reshape(np.arange(12), 2, 3, 2).MakeGeneric<int>();
+            n = np.reshape(np.arange(12), 2, 3, 2).MakeGeneric<long>();
             Assert.IsTrue(n[0, 0, 1] == 1);
             Assert.IsTrue(n[1, 0, 1] == 7);
             Assert.IsTrue(n[1, 1, 0] == 8);
 
-            n = np.reshape(np.arange(12), 3, 4).MakeGeneric<int>();
+            n = np.reshape(np.arange(12), 3, 4).MakeGeneric<long>();
             Assert.IsTrue(n[1, 1] == 5);
             Assert.IsTrue(n[2, 0] == 8);
 
-            n = np.reshape(n, 2, 6).MakeGeneric<int>();
+            n = np.reshape(n, 2, 6).MakeGeneric<long>();
 
             Assert.IsTrue(n[1, 0] == 6);
         }
@@ -64,13 +64,13 @@ namespace NumSharp.UnitTest.Manipulation
             Assert.IsTrue(nd.shape[0] == 4);
             Assert.IsTrue(nd.shape[1] == 3);
 
-            nd = np.arange(12).MakeGeneric<int>();
+            nd = np.arange(12).MakeGeneric<long>();
             nd = nd.reshape(1, 3, 4);
             nd = nd.reshape(3, -1);
             Assert.IsTrue(nd.shape[0] == 3);
             Assert.IsTrue(nd.shape[1] == 4);
 
-            nd = np.arange(100 * 100 * 3).MakeGeneric<int>();
+            nd = np.arange(100 * 100 * 3).MakeGeneric<long>();
             nd = nd.reshape(100, 100, 3);
             nd = nd.reshape(-1, 3);
             Assert.IsTrue(nd.shape[0] == 10000);
@@ -86,8 +86,8 @@ namespace NumSharp.UnitTest.Manipulation
         [Test]
         public void ValueTest()
         {
-            var x = np.arange(4).MakeGeneric<int>();
-            var y = x.reshape(2, 2).MakeGeneric<int>();
+            var x = np.arange(4).MakeGeneric<long>();
+            var y = x.reshape(2, 2).MakeGeneric<long>();
             y[0, 1] = 8;
             Assert.AreEqual(x[1], y[0, 1]);
         }
