+++ test/NumSharp.UnitTest/Math/NDArray.power.Test.cs
@@ -13,9 +13,10 @@ namespace NumSharp.UnitTest
         [Test]
         public void PowerWithSingleValue()
         {
+            // np.arange returns int64 by default (NumPy 2.x)
             var nd = np.arange(3);
 
-            var nd1 = np.power(nd, 2).MakeGeneric<int>();
+            var nd1 = np.power(nd, 2).MakeGeneric<long>();
 
             Assert.IsTrue(nd1[0] == 0);
             Assert.IsTrue(nd1[1] == 1);
